Adding design notes for JPTest project
This is a scan of some notes I made for a basic architecture for the
JPTest project. I tried to add an appropriate level of abstraction so
that I can go in now and implement just enough to get basic serial comms
from the pc so we can start using JPTest, while at the same time
sketching out an idea for future development, so that when we extend it
later, we're not redoing everything.